requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cloudinary==1.44.1
cryptography==46.0.5
defusedxml==0.7.1
dj-database-url==0.5.0
Django==6.0.3
django-allauth==0.54.0
django-cloudinary-storage==0.3.0
gunicorn==25.3.0
idna==3.11
oauthlib==3.3.1
packaging==26.0
pillow==12.1.1
psycopg2==2.9.11
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.5
stripe==14.4.1
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.12.0
mailchimp3==3.0.17
